+++ tests/package.json.test.js
@@ -1,5 +1,0 @@
-describe('package.json', () => {
-  it('doesn’t have any dependencies', () => {
-    expect(require('../package.json').dependencies).toBeUndefined();
-  });
-});